Change the param type from unsigned long to size_t

+++ ft_memset.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, unsigned long len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*s;
 	unsigned char	ch;
